+++ src/server.ts
@@ -1,13 +1,12 @@
 #!/usr/bin/env node
 import chalk from "chalk";
 import fsx from "fs-extra";
-import http from "node:http";
 import figlet from "figlet";
+import webapi from "./api/webapi.js";
 import { join, dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { JobLoader } from "./tasks/JobLoader.js";
 import { JobScheduler } from "./tasks/JobScheduler.js";
-import { ENV } from "./types/Options.types.js";
 import type { Job } from "./types/Config.types.js";
 import type { RunnerResult } from "./types/Task.types.js";
 
@@ -19,55 +18,50 @@ const appDir = join(dirname(fileURLToPath(import.meta.url)), "..");
  * Entry point to start server
  */
 export async function start() {
-   // create app setup model with defaults
-   const port = process.env[ENV.PORT] ?? 8778;
-
    // create config loader & job scheduler instance
    const jobLoader = new JobLoader();
    const jobScheduler = new JobScheduler();
 
    try {
-      // remove health check file
-      await fsx.remove("/tmp/cronops_healthy");
-
       // read app package
       const packageJSON = await fsx.readJSON(join(appDir, "package.json"));
 
       // log welcome message
       console.log(figlet.textSync("CronOps", { horizontalLayout: "fitted" }));
       console.log(chalk.cyan.bold(`\n☰ CronOps v${packageJSON.version}`) + chalk.cyan.italic(` »Omnia coniuncta sunt«`));
+      console.log(`Monitoring job configs in ${join(jobLoader.configDir, "jobs")} ...`);
 
-      // remove dry_run artifacts from previous runs
+      // remove temp dir (dry_run artifacts from previous runs)
       await fsx.emptyDir(jobScheduler.tempDir);
+
+      // init web api
+      webapi(jobScheduler);
    } catch (err) {
       console.error(`CronOps initialization error. ${err instanceof Error ? err.stack : err}`);
       console.error(`Please check environment settings.`);
       process.exit(1);
    }
 
-   jobLoader.onLoading((jobsDir: string, firstRun: boolean) => {
-      if (firstRun) console.log(`\nLoading job config from ${jobsDir} ...`);
-   });
-
    jobLoader.onLoadingError((entry: string, message: string) => {
       console.log(`🔴 Error loading job '${entry}'. ${message}`);
-      //fsx.removeSync("/tmp/cronops_healthy");
    });
 
    // called if config file has been loaded/reloaded
    jobLoader.onJobLoaded((job: Job) => {
-      jobScheduler.scheduleJob(job);
+      if (job.enabled !== false) jobScheduler.scheduleJob(job);
    });
 
-   // called if config file has been loaded/reloaded
+   // called if config file has been removed from file system
    jobLoader.onJobDeleted((jobId: string) => {
       jobScheduler.unscheduleJob(jobId);
-      console.log(`⭕ job [${jobId}] unscheduled (removed from file system)`);
    });
 
-   jobScheduler.onJobScheduled((job: Job, reschedule) => {
-      if (job.enabled === false) console.log(`⭕ job [${job.id}] unscheduled (disabled)`);
-      else console.log(`🕔 job [${job.id}] ${reschedule ? "re-" : ""}scheduled (${chalk.greenBright(job.cron)})${job.dry_run ? " 👋 DRY-RUN mode!" : ""}`);
+   jobScheduler.onChanged((isReload: boolean) => {
+      const jobs = jobScheduler.getScheduledJobs();
+      console.log(`\nJob config ${isReload ? "changed" : "loaded"} (${plural(jobs.length, "active job")})`);
+      for (const job of jobs) {
+         console.log(` 🕔 [${job.id}] scheduled (${chalk.greenBright(job.cron)})${job.dry_run ? " 👋 DRY-RUN mode!" : ""}`);
+      }
    });
 
    jobScheduler.onJobError((job: Job, err: Error) => {
@@ -108,38 +102,9 @@ export async function start() {
       process.exit(-1);
    });
 
-   // create simple http admin endpoint to manually trigger jobs via HTTP POST /trigger/{jobId}
-   const httpServer = http.createServer(async (req, res) => {
-      const { method, url } = req;
-
-      // Pattern Matching für /trigger/{{jobId}}
-      if (method === "POST" && url?.startsWith("/trigger/")) {
-         const jobId = url.split("/")[2] || "";
-         if (!jobScheduler.isJobScheduled(jobId)) return res.writeHead(400).end(`Job '${jobId}' not found\n`);
-         try {
-            console.log(`[${jobId}] JOB manually triggered via HTTP request`);
-            jobScheduler.executeJob(jobId);
-            res.writeHead(202, { "Content-Type": "application/json" });
-            return res.end(JSON.stringify({ accepted: true, job: jobId }, null, 3));
-         } catch (error) {
-            console.error(`[${jobId}] ${String(error)}`);
-            res.writeHead(202, { "Content-Type": "application/json" });
-            return res.end(JSON.stringify({ accepted: true, job: jobId }, null, 3));
-         }
-      }
-      return res.writeHead(404).end("Not found!");
-   });
-
-   // start http server
-   httpServer.listen(port, () => {
-      console.log(`Web-Admin API listening on port ${port} ...`);
-      console.log(chalk.gray(`⎆ to get server status, type           curl -X GET http://localhost:${port}/status`));
-      console.log(chalk.gray(`⎆ to trigger a job manually, type      curl -X POST http://localhost:${port}/trigger/{job-id}`));
-      console.log(chalk.gray(`⎆ to gracefully terminate server,type  curl -X POST http://localhost:${port}/terminate`));
-   });
-
    // schedule config loader, load config & schedule jobs
    jobLoader.schedule(true);
+   jobScheduler.schedule();
 }
 
 // server entry point