+++ src/tasks/JobLoader.ts
@@ -1,7 +1,7 @@
 import fsx from "fs-extra";
 import YAML from "yaml";
 import glob from "fast-glob";
-import { join, dirname, parse, resolve } from "node:path";
+import { join, dirname, resolve, basename } from "node:path";
 import { AbstractTask } from "./AbstractTask.js";
 import { ZodError } from "zod";
 import { fileURLToPath } from "node:url";
@@ -11,48 +11,48 @@ import { type Job, JobSchema } from "../types/Config.types.js";
 import type { FileHistory } from "../types/Task.types.js";
 
 const appDir = join(dirname(fileURLToPath(import.meta.url)), "..", "..");
+const entry2id = (str: string) => join(dirname(str), basename(str, ".yaml"));
 
 export class JobLoader extends AbstractTask<Job[]> {
+   public configDir: string;
    private firstRun = true;
-   private jobsDir: string;
    private jobHistory: FileHistory;
 
    constructor(options: LoaderOptions = {}) {
       super("*/8 * * * * *");
-      this.jobsDir = resolve(options.configDir ?? process.env[ENV.CONFIG_DIR] ?? "./config", "jobs");
+      this.configDir = resolve(options.configDir ?? process.env[ENV.CONFIG_DIR] ?? "./config");
       this.jobHistory = new FileHistoryModel();
    }
 
    protected override async run(): Promise<Job[]> {
       const result: Job[] = [];
       const ttime = Date.now();
+      const jobsDir = join(this.configDir, "jobs");
 
-      if (this.firstRun && !fsx.pathExistsSync(this.jobsDir))
+      if (this.firstRun && !fsx.pathExistsSync(jobsDir))
          // does config exist on first start?
          try {
-            // try copy default config
-            await fsx.copy(join(appDir, "config", "jobs"), this.jobsDir);
+            // copy default config
+            await fsx.copy(join(appDir, "config"), this.configDir);
          } catch {
             // nop
          }
 
       // scan job config files
-      const entries = await glob(["**/*.yaml"], { cwd: this.jobsDir });
+      const entries = await glob(["**/*.yaml"], { cwd: jobsDir });
 
       // notify listeners
-      this.events.emit("loading", this.jobsDir, this.firstRun);
       this.firstRun = false;
 
       // loop all found job entries
       for (const entry of entries) {
-         const { name: id } = parse(entry);
-         const jobFile = join(this.jobsDir, entry);
+         const jobFile = join(jobsDir, entry);
          try {
             const stats = await fsx.stat(jobFile);
-            const { changed, added } = this.jobHistory.updateSourceEntry(jobFile, [stats.mtimeMs, ttime]);
+            const { changed, added } = this.jobHistory.updateSourceEntry(entry, [stats.mtimeMs, ttime]);
             if (changed) {
                const jobConfig = JobSchema.parse(YAML.parse(await fsx.readFile(jobFile, "utf-8")));
-               const job = { id, ...jobConfig } as Job;
+               const job = { id: entry2id(entry), ...jobConfig } as Job;
                result.push(job);
                this.events.emit("job-loaded", job, !added);
             }
@@ -66,12 +66,14 @@ export class JobLoader extends AbstractTask<Job[]> {
       const removedJobs = this.jobHistory.cleanup();
 
       // remove jobs
-      for (const jobFile of removedJobs) {
-         const { name } = parse(jobFile);
-         this.events.emit("job-deleted", name);
+      for (const entry of removedJobs) {
+         this.events.emit("job-deleted", entry2id(entry));
       }
 
-      // return JobConfig or undefined (no action)
+      // notify loaded listener
+      this.events.emit("loaded", result);
+
+      // return (re)loaded jobs
       return result;
    }
 
@@ -79,8 +81,8 @@ export class JobLoader extends AbstractTask<Job[]> {
       return await this.run();
    }
 
-   public onLoading(cb: (jobsDir: string, firstRun: boolean) => void) {
-      this.events.on("loading", cb);
+   public onceLoaded(cb: (jobs: Job[]) => void) {
+      this.events.on("loaded", cb);
    }
 
    public onLoadingError(cb: (jobId: string, message: string) => void) {