+++ src/tasks/JobScheduler.ts
@@ -1,18 +1,20 @@
-import { EventEmitter } from "node:events";
 import { JobRunner } from "./JobRunner.js";
 import { JobModel } from "../models/JobModel.js";
 import { JobRunnerSetup } from "../models/JobRunnerSetup.js";
+import { AbstractTask } from "./AbstractTask.js";
 import type { Job } from "../types/Config.types.js";
 import type { RunnerResult } from "../types/Task.types.js";
 import type { RunnerOptions } from "../types/Options.types.js";
+import { JobError } from "../errors/JobError.js";
 
-export class JobScheduler {
-   protected events = new EventEmitter();
+export class JobScheduler extends AbstractTask<void> {
    protected runnerSetup: JobRunnerSetup;
    protected runnerMap: Map<string, JobRunner>;
-   protected rescheduled = false;
+   private changed = false;
+   private isReload = false;
 
    constructor(options: RunnerOptions = {}) {
+      super("*/5 * * * * *");
       this.runnerSetup = new JobRunnerSetup(options);
       this.runnerMap = new Map();
    }
@@ -25,23 +27,29 @@ export class JobScheduler {
       return this.runnerSetup.tempDir;
    }
 
-   public unscheduleAll() {
-      for (const runner of this.runnerMap.values()) runner.unschedule();
-      this.runnerMap = new Map();
+   protected override async run(): Promise<void> {
+      if (this.changed) {
+         this.changed = false;
+         this.events.emit("schedule-changed", this.isReload);
+         this.isReload = true;
+      }
    }
 
-   public scheduleJobs(jobs: Job[]) {
+   public unscheduleAll() {
+      for (const runner of this.runnerMap.values()) runner.unschedule();
+      this.runnerMap = new Map<string, JobRunner>();
+      this.changed = true;
+   }
+
+   public scheduleJobs(jobs: Job[], cb?: (count: number) => void) {
       // unschedule all scheduled job tasks in the background
       this.unscheduleAll();
 
-      // schedule all configured jobs
+      // schedule jobs
       for (const jobEntry of jobs) this.scheduleJob(jobEntry);
 
-      // notify listener if no jobs
-      this.events.emit("ready", this.runnerMap.size, this.rescheduled);
-
-      // mark next run as "reschedule"
-      this.rescheduled = true;
+      // callback
+      if (cb) cb(this.runnerMap.size);
    }
 
    /**
@@ -49,12 +57,15 @@ export class JobScheduler {
     * @param job job definition
     * @param defaults defaults to use if not defined in job definition
     * @throws JobError if job definition is not valid
-    * @throws ValidationError if Job config is not valid
+    * @throws JobError if Job config is not valid
     */
    public scheduleJob(job: Job, defaults = {}) {
       let rescheduled = false;
 
-      // validate job -> throws ValidationError
+      // throw error if job is disabled
+      if (job.enabled === false) JobError.throw(job.id, `Cannot schedule disabled job [${job.id}]!`);
+
+      // validate job --> throws JobError
       this.runnerSetup.validateJob(job);
 
       // unschedule already existing job
@@ -76,12 +87,16 @@ export class JobScheduler {
 
       // start cron scheduler for this job
       task.schedule();
+
+      // mark as changed
+      this.changed = true;
    }
 
    public unscheduleJob(jobId: string) {
       const task = this.runnerMap.get(jobId);
       if (task) {
          this.runnerMap.delete(jobId);
+         this.changed = true;
          task.unschedule();
       }
    }
@@ -99,18 +114,24 @@ export class JobScheduler {
       this.runnerSetup.validateJob(job);
    }
 
-   public async gracefulTerminate(timeout: number = 1000) {
+   public getScheduledJobs(): Job[] {
+      return Array.from(this.runnerMap.values()).map((runner) => runner.job);
+   }
+
+   public override async gracefulTerminate(timeout: number = 1000) {
       const jobRunnerTasks = [...this.runnerMap.values()];
       this.unscheduleAll();
+      // graceful terminate all scheduled tasks including "self"
+      await super.gracefulTerminate(timeout);
       for (const task of jobRunnerTasks) await task.gracefulTerminate(timeout);
    }
 
    /**
-    * Emitted if all configured jobs are initialized & scheduled.
-    * @param cb
+    * Emitted if job schedule has been changed (checked every 5 seconds)
+    * @param cb callback function
     */
-   public onReady(cb: (jobCount: number, rescheduled: boolean) => void) {
-      this.events.on("ready", cb);
+   public onChanged(cb: (initialConfig: boolean) => void) {
+      this.events.on("schedule-changed", cb);
    }
 
    /**
@@ -138,7 +159,7 @@ export class JobScheduler {
    }
 
    /**
-    * Emitted if there was any kind of activity on the file system during job execution.
+    * Emitted if there was any kind of activity during job execution.
     * The `activity` parameter can have the followings values:
     * - `EXECUTED`: a command has been executed (globally or on a selected source file)
     * - `COPIED`: a source file has been copied to `path`