+++ src/tests/loadtest.ts
@@ -1,17 +1,19 @@
 import fsx from "fs-extra";
-import { createJobScheduler } from "../index.js";
+import { createJobScheduler, type Job } from "../index.js";
 
 let time = Date.now();
 
 /**
- * LOADTEST
+ * CronOps LOADTEST
  *
  * Performs 5 jobs on ALL FILES in the node_modules folder of this projects (~10k files):
  * - COPY all files from ./node_modules to ./build/loadtest/copied
  * - MOVE all files from ./build/loadtest/copied to ./build/loadtest/moved
- * - EXECUTE `chmod ugo+rwx` on all `*.md`` files in ./build/loadtest/moved
+ * - EXECUTE `chmod ugo+rwx` on all `*.ts`` files in ./build/loadtest/moved
  * - ARCHIVE all files in ./build/loadtest/moved to ./build/loadtest/loadtest.tgz
  * - DELETE all files in ./build/loadtest/moved
+ *
+ * Use `npm run loadtest` to execute load test.
  */
 
 // init
@@ -19,37 +21,7 @@ await fsx.emptyDir("./build/loadtest");
 await fsx.ensureDir("./build/loadtest/copied");
 await fsx.ensureDir("./build/loadtest/moved");
 
-// create job scheduler instance
-const scheduler = createJobScheduler({
-   logDir: "./build/loadtest",
-});
-
-scheduler.onReady(() => {
-   // begin manual execution ...
-   scheduler.executeJob("copy");
-});
-
-scheduler.onJobStarted((job) => {
-   console.log(`Starting job [${job.id}] ...`);
-   time = Date.now();
-});
-
-scheduler.onJobFinished((job, { copied, deleted, archived, executed }) => {
-   console.log(`✔ Done in ${Date.now() - time}ms (${copied} copied, ${deleted} deleted, ${archived} archived, ${executed} executed)`);
-   // manually start next jobs
-   if (job.id === "copy") scheduler.executeJob("move");
-   else if (job.id === "move") scheduler.executeJob("execute");
-   else if (job.id === "execute") scheduler.executeJob("archive");
-   else if (job.id === "archive") scheduler.executeJob("delete");
-   else done();
-});
-
-scheduler.onJobError((job, err) => {
-   console.log(`${job.id} failed. ${String(err)}`);
-});
-
-// finally schedule the 4 jobs
-scheduler.scheduleJobs([
+const jobs: Job[] = [
    {
       id: "copy",
       action: "copy",
@@ -110,7 +82,37 @@ scheduler.scheduleJobs([
          includes: ["**/**"],
       },
    },
-]);
+];
+
+// create job scheduler instance
+const scheduler = createJobScheduler({
+   logDir: "./build/loadtest",
+});
+
+scheduler.onJobStarted((job) => {
+   console.log(`Starting job [${job.id}] ...`);
+   time = Date.now();
+});
+
+scheduler.onJobFinished((job, { copied, deleted, archived, executed }) => {
+   console.log(`✔ Done in ${Date.now() - time}ms (${copied} copied, ${deleted} deleted, ${archived} archived, ${executed} executed)`);
+   // manually start next jobs
+   if (job.id === "copy") scheduler.executeJob("move");
+   else if (job.id === "move") scheduler.executeJob("execute");
+   else if (job.id === "execute") scheduler.executeJob("archive");
+   else if (job.id === "archive") scheduler.executeJob("delete");
+   else done();
+});
+
+scheduler.onJobError((job, err) => {
+   console.log(`${job.id} failed. ${String(err)}`);
+});
+
+// schedule all jobs jobs and start with copy job
+scheduler.scheduleJobs(jobs, () => {
+   // begin manual execution ...
+   scheduler.executeJob("copy");
+});
 
 async function done() {
    // should throw error if dir is not empty (= test failed!)